requirements.txt
Django==6.0.3
psycopg2-binary==2.9.11
djangorestframework==3.16.1
django-environ==0.13.0
django-simple-history==3.11.0
django-cors-headers==4.9.0
django-jazzmin==3.0.4
openpyxl==3.1.2
